Extract main.py
