+++ setup.py
@@ -30,7 +30,7 @@ else:
             ])
     extra['entry_points'] = {
         'console_scripts': [
-            'batch_scoring = datarobot_batch_scoring.batch_scoring:main']}
+            'batch_scoring = datarobot_batch_scoring.main:main']}
     extra['install_requires'] = install_requires
